fix: refactor storagePlugins
